requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
lxml==4.6.3
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
